Changed location of image_mute_radar
