+++ pyart/util/__init__.py
@@ -15,10 +15,9 @@ from .datetime_utils import datetime_from_dataset, datetimes_from_dataset
 from .datetime_utils import datetime_from_grid
 from .xsect import cross_section_ppi, cross_section_rhi
 from .hildebrand_sekhon import estimate_noise_hs74
-from .radar_utils import is_vpt, to_vpt, join_radar
+from .radar_utils import is_vpt, to_vpt, join_radar, image_mute_radar
 from .simulated_vel import simulated_vel_from_profile
 from .sigmath import texture_along_ray, rolling_window
 from .sigmath import texture, angular_texture_2d
-from .image_mute import image_mute_radar
 
 __all__ = [s for s in dir() if not s.startswith('_')]
